% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geo_unit.R
\name{add_geo_unit}
\alias{add_geo_unit}
\title{Add geoid to survey data}
\usage{
add_geo_unit(
  target_df,
  input_geo_filepath,
  geotype,
  lat_long_columns = c("long", "lat"),
  input_crs = 4326,
  target_crs = 4269
)
}
\arguments{
\item{target_df}{The target dataframe with coordinates. Usually, the individual_repondent table in the YPCCC database.}

\item{input_geo_filepath}{Path to the .rda file of the geographic data}

\item{geotype}{Type of the geographic data. Possible values: "state", "county", "tract", "cd116", "sldl", "sldu", etc.}

\item{lat_long_columns}{Column names for longitude and latitude in the target_df. Default = c("long", "lat")}

\item{input_crs}{Coordinate system of the survey coordinates. Default = EPSG 4326}

\item{target_crs}{Coordinate system of the geograohic data. Default = EPSG 4269}
}
\value{
Tibble dataframe with a column for the geographic unit
}
\description{
Perform spatial join between respondent's coordinates and the exisitng geographic data of the US
}
