% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rows_to_table.R
\name{add_rows_to_table}
\alias{add_rows_to_table}
\title{Add dataframe rows to a database table}
\usage{
add_rows_to_table(con, db_table, db_columns, df_name)
}
\arguments{
\item{con}{Database connection created using connect_db() or DBI::dbConnect()}

\item{db_table}{Character. Name of the table to be updated within the database}

\item{db_columns}{Character. List of column names of the table to be updated within the database. The format should in be an SQL syntax: (col1, col2, col3, ...)}

\item{df_name}{Dataframe with the updated values}
}
\value{
Updated dataframe from the database
}
\description{
Takes a database connection with a table and add to it the selected column values from a dataframe
}
\examples{
add_rows_to_table(con, 'survey', '(source, wave, year)', survey)

}
