% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to an existing database}
\usage{
connect_db(dbname, host, username, password, local_infile = NULL)
}
\arguments{
\item{dbname}{Character. Name of the database schema.}

\item{host}{Character. Name of the host.}

\item{username}{Character. Username.}

\item{password}{Character. Password.}

\item{local_infile}{Boolean. Set to TRUE when necessary.}
}
\value{
Database connection
}
\description{
Takes username and password to connect to an existing database, which will allow the user to query or updata the database. This can also be done using DBI::dbConnect()
}
\examples{
connect_db('testdb', 'localhost', 'root', 'password')

}
