% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_schema.R
\name{update_schema}
\alias{update_schema}
\title{Update the structure of an existing database}
\usage{
update_schema(con, sql_file)
}
\arguments{
\item{con}{Database connection}

\item{sql_file}{Path to .sql file from MySQL}
}
\value{
None
}
\description{
Use an existing SQL file containing statement about the schema to update the database structure. The SQL file can be created using MySQL Workbench by exporting the database.
}
\examples{
update_schema(con, '/Users/user/Desktop/structure.sql')

}
